Bungee stuff, Fixed bugs

+++ FWCore/src/net/clashwars/cwcore/commands/KillCmd.java
@@ -1,10 +1,14 @@
 package net.clashwars.cwcore.commands;
 
+import java.io.ByteArrayOutputStream;
+import java.io.DataOutputStream;
+
 import net.clashwars.cwcore.CWCore;
 import net.clashwars.cwcore.commands.internal.CommandClass;
 import net.clashwars.cwcore.entity.CWPlayer;
 import net.clashwars.cwcore.util.CmdUtils;
 
+import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
@@ -22,6 +26,7 @@ public class KillCmd implements CommandClass {
 	public boolean execute(CommandSender sender, Command cmd, String lbl, String[] args) {
 		String pf = cwc.getPrefix();
 		Player player = null;
+		String pplayer = null;
 		CWPlayer cwp = null;
 		
 		/* Modifiers + No args */
@@ -32,6 +37,7 @@ public class KillCmd implements CommandClass {
 			sender.sendMessage(pf + "Modifiers: ");
 			sender.sendMessage(ChatColor.DARK_PURPLE + "-s" + ChatColor.DARK_GRAY + " - " + ChatColor.GRAY + "No messages");
 			sender.sendMessage(ChatColor.DARK_PURPLE + "-f" + ChatColor.DARK_GRAY + " - " + ChatColor.GRAY + "Force kill");
+			sender.sendMessage(ChatColor.DARK_PURPLE + "-*" + ChatColor.DARK_GRAY + " - " + ChatColor.GRAY + "Kill players on other servers.");
 			return true;
 		}
 		boolean silent = false;
@@ -44,32 +50,56 @@ public class KillCmd implements CommandClass {
 			force = true;
 			args = CmdUtils.modifiedArgs(args,"-f", true);
 		}
+		boolean bungee = false;
+		if (CmdUtils.hasModifier(args,"-*", true)) {
+			bungee = true;
+			args = CmdUtils.modifiedArgs(args,"-*", true);
+		}
 		
 		/* 1 arg (Player) */
 		if (args.length >= 1) {
 			player = cwc.getServer().getPlayer(args[0]);
+			pplayer = args[0];
 		}
 		
 		/* null checks */
-		if (player == null) {
-			sender.sendMessage(pf + ChatColor.RED + "Invalid player.");
-			return true;
-		}
-		cwp = cwc.getPlayerManager().getPlayer(player);
-		if (!force) {
-			if (cwp.getGamemode() == 1 || cwp.getGod() == true) {
-				sender.sendMessage(pf + ChatColor.RED + "You can't kill this player.");
+		if (!bungee) { 
+			if (player == null) {
+				sender.sendMessage(pf + ChatColor.RED + "Invalid player.");
 				return true;
+			}
+			cwp = cwc.getPlayerManager().getPlayer(player);
+			if (!force) {
+				if (cwp.getGamemode() == 1 || cwp.getGod() == true) {
+					sender.sendMessage(pf + ChatColor.RED + "You can't kill this player.");
+					return true;
+				}
 			}
 		}
 			
-		
 		/* Action */
-		player.setHealth(0);
-		if (!silent) {
-			player.sendMessage(pf + "You where killed by " + ChatColor.DARK_PURPLE + sender.getName());
-			if (sender.getName() != player.getName()) {
-				sender.sendMessage(pf + "You have killed " + ChatColor.DARK_PURPLE + player.getDisplayName());
+		if (bungee) {
+			try {
+				ByteArrayOutputStream b = new ByteArrayOutputStream();
+				DataOutputStream out = new DataOutputStream(b);
+
+				out.writeUTF("Kill");
+				out.writeUTF(sender.getName());
+				out.writeUTF(pplayer);
+				out.writeBoolean(silent);
+				out.writeBoolean(force);
+
+				Bukkit.getOnlinePlayers()[0].sendPluginMessage(cwc.getPlugin(), "CWBungee", b.toByteArray());
+			} catch (Throwable e) {
+				e.printStackTrace();
+			}
+		} else {
+			player.setHealth(0);
+			if (!silent) {
+				player.sendMessage(pf + "You where killed by " + ChatColor.DARK_PURPLE + sender.getName());
+				if (sender.getName() != player.getName()) {
+					sender.sendMessage(pf + "You have killed " + ChatColor.DARK_PURPLE + player.getDisplayName());
+				}
 			}
 		}
 		return true;
